修复 courier 编译错误：KDoc嵌套注释/waitFor类型/流读取超时常量/命名参数对齐

+++ app/src/main/java/com/newoether/agora/courier/CourierFileAccess.kt
@@ -218,7 +218,7 @@ class CourierFileAccess(
                 val process = ProcessBuilder("su", "-c", "id")
                     .redirectErrorStream(true)
                     .start()
-                val done = process.waitFor(Constants.COURIER_SU_PROBE_TIMEOUT_MS, TimeUnit.MILLISECONDS)
+                val done = process.waitFor(Constants.COURIER_SU_PROBE_TIMEOUT_MS.toLong(), TimeUnit.MILLISECONDS)
                 process.inputStream.use { it.readBytes() } // drain to avoid buffer deadlock
                 val ok = done && process.exitValue() == 0
                 if (!done) process.destroy()
